Added user position
For geolocation to work, the browser needs to allow the local server access to the location. Because the connection isn't secure, it is blocked by default for security reasons. You can disable it like so: https://stackoverflow.com/questions/57450093/browser-says-camera-blocked-to-protect-your-privacy (This says for the camera, but it applies to all browser permissions.) After that you can click the button on the top right to give the map your location.

The user location is now the start coordinate and the end coordinate is now the most recent click on the map. User location and end coordinate are fed to shortest route.